moved Button.jsx to a separate folder, added Button.scss and moved all styles to it
